requirements.txt
streamlit==1.38.0
streamlit-extras==0.4.7
markdown2==2.5.0
python-frontmatter==1.0.0
Flask==3.0.3
json-logging==1.3.0
python-dotenv==1.0.1
